# Enable cudagraphs compile backend