--- a/model.py
+++ b/model.py
@@ -81,11 +81,8 @@ class LayerStacks(nn.Module):
         self.output.bias = nn.Parameter(output_bias)
 
     def forward(self, x, ls_indices):
-        # Precompute and cache the offset for gathers
-        if self.idx_offset == None or self.idx_offset.shape[0] != x.shape[0]:
-            self.idx_offset = torch.arange(
-                0, x.shape[0] * self.count, self.count, device=ls_indices.device
-            )
+
+        assert self.idx_offset != None and self.idx_offset.shape[0] == x.shape[0]
 
         indices = ls_indices.flatten() + self.idx_offset