@Disservin 2025-06-28
i also get a speedup for the smallnet +~12 it/s and a slowdown for the bignet