@vondele 2025-06-28
for the bignet, you mean slowdown when you use cudagraphs or for the branch? (It should be essentially equivalent without the option).